refactor notify court email
