+++ apps/plea/tasks.py
@@ -64,12 +64,6 @@ def email_send_court(self, case_id, count_id, email_data):
     case = Case.objects.get(pk=case_id)
 
     court_obj = get_court(email_data["case"]["urn"], case.ou_code)
-    print(f'COURT ID and SUB EMAIL: {court_obj.id, court_obj.submission_email}'
-          f'COURT NAME: {court_obj.court_name}'
-          f'\n\n\n', flush=True)
-    logger.info(f'COURT ID and SUB EMAIL: {court_obj.id, court_obj.submission_email}'
-                + 'COURT NAME: {court_obj.court_name}'
-                + '\n\n\n')
 
     plea_email_to = court_obj.submission_email
     # smtp_route = get_smtp_gateway(court_obj.submission_email)
@@ -86,7 +80,6 @@ def email_send_court(self, case_id, count_id, email_data):
     personalisation = {
         "case_id": case.id,
         "count_id": count_id,
-        "link_to_file": "",
         "subject": email_subject
     }
     plea_email = GovNotify(
